Moved location client to dashboard
